0_Azure/3_AzureAI/14_AIUseCases/0_PDFProcessingAnalytics/src/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pdfminer.six
azure-cosmos==4.3.0